cpplint: overall cleanup

+++ benchmarks/groupby/groupby.cpp
@@ -76,7 +76,7 @@ int main(int argc, char* argv[]) {
                           LOG1 //<< "\n"
                               << "RESULT"
                               << " name=total"
-                              << " time=" << (double)timer.Milliseconds()
+                              << " time=" << timer.Milliseconds()
                               << " filename=" << input
                               << " sanity1=" << res1
                               << " sanity2=" << res2;
